+++ ChangeLog.md
@@ -173,6 +173,7 @@
 
 ###### Issues fixed
 - [`KT-9156`](https://youtrack.jetbrains.com/issue/KT-9156) Quote non-identifier names in Kotlin references
+- [`KT-9157`](https://youtrack.jetbrains.com/issue/KT-9157) Fixed in-place rename of Kotlin expression referring Java declaration
 
 #### Java to Kotlin converter